remove java.lang.Double usages

+++ src/jvmMain/kotlin/net/dinkla/raytracer/objects/acceleration/kdtree/builder/Test2Builder.kt
@@ -107,7 +107,7 @@ class Test2Builder : TreeBuilder {
 
                 fun max(): Split {
                     val s = Split(null)
-                    s.sah = java.lang.Float.POSITIVE_INFINITY.toDouble()
+                    s.sah = Double.POSITIVE_INFINITY
                     return s
                 }
             }
